+++ src/patterns/SaleCard/SaleCard.stories.tsx
@@ -6,8 +6,8 @@ import { Icon } from '../../components/Icon';
 import { getScssVar } from '../../utils/scssUtils';
 import SaleCard, { SaleCardProps } from './SaleCard';
 import { SaleCardActions } from './SaleCardActions';
-import { SaleCardVariants } from './types';
 import { Text } from '../../components/Text';
+import { SaleCardImageDisplay, SaleCardVariants } from './types';
 
 const meta = {
   title: 'Patterns/SaleCard',
@@ -266,4 +266,32 @@ export const SaleCardNoImage = () => (
     location="New York"
     badgeText="Happening Now"
   />
+);
+
+export const SaleCardImageRightVariants = () => (
+  <div style={{ display: 'flex', gap: 'var(--spacing-sm)', flexDirection: 'column' }}>
+    <SaleCard
+      {...defaultProps}
+      variant={SaleCardVariants.DEFAULT}
+      imageDisplay={SaleCardImageDisplay.RIGHT}
+      badgeText="Happening Now"
+    />
+    {divider}
+    <SaleCard
+      {...defaultProps}
+      variant={SaleCardVariants.RELATED_SALE_TILE}
+      imageDisplay={SaleCardImageDisplay.RIGHT}
+      badgeText="Accepting Consignments"
+      imageSrc="https://www.assets.phillips.com/image/upload/t_Website_AuctionPageHero/v1/QA/auctions/NY041325/NY041325.jpg"
+    />
+    {divider}
+    <SaleCard
+      {...defaultProps}
+      variant={SaleCardVariants.RELATED_SALE_TILE}
+      imageDisplay={SaleCardImageDisplay.RIGHT}
+      titleText="Modern & Contemporary: Online Auction, New York"
+      date="27 Aug - 05 Sep"
+      badgeText="Bidding Open"
+    />
+  </div>
 );